requirements.txt
youtube_transcript_api
google-generativeai
streamlit
python-dotenv
pathlib